vm: make NewBigInteger accept int64
It is more correct upconvert to int64 if needed,
so precision is never lost.

+++ pkg/vm/stack_item.go
@@ -191,9 +191,9 @@ type BigIntegerItem struct {
 }
 
 // NewBigIntegerItem returns an new BigIntegerItem object.
-func NewBigIntegerItem(value int) *BigIntegerItem {
+func NewBigIntegerItem(value int64) *BigIntegerItem {
 	return &BigIntegerItem{
-		value: big.NewInt(int64(value)),
+		value: big.NewInt(value),
 	}
 }
